% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogemetric-package.R
\docType{package}
\name{phylogemetric-package}
\alias{phylogemetric}
\alias{phylogemetric-package}
\title{phylogemetric: Calculate Delta and Q-residual Scores}
\description{
A R library for calculating delta score (Holland et al. 2002) and Q-Residual (Gray et al. 2010) for phylogenetic data. Based on the python package phylogemetric (https://github.com/SimonGreenhill/phylogemetric)
}
\keyword{internal}
