% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{quartet_score}
\alias{quartet_score}
\title{Title}
\usage{
quartet_score(data, quartet, method = "hamming")
}
\arguments{
\item{data}{the data matrix that will be used to calculate distances}

\item{quartet}{the quartets that will be analysed. Can be derived from `quartet_sets`}

\item{method}{}
}
\value{
The quartet analysed and the delta score for that quartet
}
\description{
Title
}
