% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{quartet_sets}
\alias{quartet_sets}
\title{Title}
\usage{
quartet_sets(taxa, k = 4)
}
\arguments{
\item{taxa}{A vector of the taxa used}

\item{k}{indicates the set size. Defaults to 4. You are unlikely to need to change this}
}
\value{
The number of k size sets
}
\description{
Title
}
