% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{summarise_taxon_scores}
\alias{summarise_taxon_scores}
\title{Title}
\usage{
summarise_taxon_scores(sets, quartet_distances)
}
\arguments{
\item{sets}{The quartet sets that will be analysed}

\item{quartet_distances}{The quartet delta scores that were output from `quartet_score`}
}
\description{
Title
}
